requirements.txt
requests==2.23.0
ptvsd==4.3.2 # Required for debugging.
astroid==2.4.2
click==7.1.2
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.0
Pillow==7.2.0
pylint==2.5.3
six==1.15.0
toml==0.10.1
torch==1.5.1
torchvision==0.2.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1